Move Private class implementation into own CPP file (Head)
